[LLDB][NVIDIA] Add support for reading all missing registers
This implementation reads all the uniform registers for each thread, but
I'll fix that in my next patches.

+++ lldb/tools/lldb-server/Plugins/NVIDIAGPU/RegisterContextNVIDIAGPU.h
@@ -32,10 +32,10 @@ struct ThreadRegistersValues {
   uint64_t errorPC;
   uint32_t R[kNumRRegs];
   uint32_t RZ;             // R255 - zero register
-  uint8_t P[kNumPRegs];    // Predicate registers (1-bit each, stored in bytes)
+  uint32_t P[kNumPRegs];   // Predicate registers (1-bit each, stored in bytes)
   uint32_t UR[kNumURRegs]; // Uniform registers
   uint32_t URZ;            // UR255 - uniform zero register
-  uint8_t UP[kNumUPRegs];  // Uniform predicate registers (1-bit each, stored in
+  uint32_t UP[kNumUPRegs]; // Uniform predicate registers (1-bit each, stored in
                            // bytes)
 };
